phase_03/backend/requirements.txt
# Plannior Phase III - Backend Dependencies
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlmodel==0.0.14
sqlalchemy==2.0.23
psycopg[binary]==3.9.1
alembic==1.12.1

# AI/LLM Integration
openai==1.3.9
tiktoken==0.5.2

# Security & JWT
PyJWT==2.8.1
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# MCP SDK (Official)
# mcp>=0.1.0  # Install when official MCP SDK is available

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Code Quality & Linting
black==23.12.0
pylint==3.0.3
flake8==6.1.0
mypy==1.7.1

# Development
ipython==8.18.1
